simplifying big assets sidebar component
